Add language update during limesurvey transformations.
The language in the translated fields is updated post
limesurvey convertion, this was made because updating the
language on each of the filters could be a more complicated
task.

Since the mark up for the translations is the same updating
the language after the limesurvey content has been extracted
produces the same output.