Add support to use extracted localised settings.
The name and description of the survey can be extracted
using the form tree.

The initial page is harcoded since it seems to be the same
disclaimer for every survey:

"The views expressed in this public consultation document
may not be interpreted as stating an official position
of the European Commission.  All definitions provided in
this document are strictly for the purposes of this public
consultation and are without prejudice to differing definitions
the Commission may use under current or future EU law, including
any revision of the definitions by the Commission concerning
the same subject matters."